break and continue.rs
When the break statement is encountered inside a loop, the loop is immediately terminated, and program control resumes at the next statement following the loop. 
And when a continue statement is encountered inside a loop, control jumps to the beginning of the loop for next iteration, skipping the execution of statements inside the body of loop for the current iteration.